Move DeploymentInfo and DeploymentMetadata to deployment package.
